+++ src/chipshot/reader/__init__.py
@@ -8,8 +8,9 @@ import logging
 import pathlib
 import typing as t
 
+from ..config import get_config_value
 from ..shared import FileInfo
-from . import encoding, header, newlines, prolog
+from . import encoding, header, newlines, prologue
 
 log = logging.getLogger(__name__)
 
@@ -30,7 +31,7 @@ def read(path: pathlib.Path, config: dict[str, t.Any]) -> FileInfo:
 
     encoding.handle(info)
     newlines.handle(info)
-    prolog.handle(info, config)
+    prologue.handle(info, config)
     header.handle(info, config)
 
     return info
@@ -39,15 +40,5 @@ def read(path: pathlib.Path, config: dict[str, t.Any]) -> FileInfo:
 def _determine_default_encoding(path: pathlib.Path, config: dict[str, t.Any]) -> str:
     """Determine the default encoding for the given path."""
 
-    suffixes = "".join(path.suffixes)[1:]
-    suffix = "".join(path.suffix)[1:]
-
-    suffixes_config = config["extension"].get(suffixes, {})
-    suffix_config = config["extension"].get(suffix, {})
-
-    default_encoding = str(
-        suffixes_config.get("encoding")
-        or suffix_config.get("encoding")
-        or config["default"]["file_info"]["encoding"]
-    )
-    return default_encoding
+    (default_encoding,) = get_config_value(config, path, "encoding")
+    return str(default_encoding)